Completed s=launch-script

+++ launch.sh
@@ -1,7 +1,11 @@
 ###################### Ilyasse Fakhreddine #######################
 
-
 #!/bin/bash
+
+if [[ $EUID -eq 0 ]]; then
+    echo "Please do not run this as root."
+    exit
+fi
 
 trap "kill 0" SIGINT
 
@@ -19,31 +23,31 @@ sleep 1
 
 echo "Starting roscore, yarpserver and iCub Simulation nodes..."
 
-roscore >/dev/null &
+gnome-terminal --tab -- "roscore"
 sleep 2
-sudo yarpserver >/dev/null &
+gnome-terminal --tab -- bash -c "echo '123' | sudo -S yarpserver"
+gnome-terminal --tab -- "iCub_SIM"
 sleep 2
-iCub_SIM >/dev/null &
+gnome-terminal -- "`pwd`/icub-control/ice_controller"
+gnome-terminal --tab -- bash -c "echo 'Go back to the first tab.'; read -n 1 -s"
 
 echo "roscore... Ok"
-echo "yarpserver... Ok"
 echo "icubsim... Ok"
-
-sleep 2
-
 echo "Starting robot controller..."
-
-gnome-terminal -e "`pwd`/icub-control/icub-control" &
-echo "Node icub-controller... Ok"
+echo "Node ice_controller... Ok"
 
 sleep 2
 
-echo "setup completed... Press any key to start the RAT Model simulation."
+echo "========================================================================="
+echo "||                                                                     ||"
+echo "|| Setup completed... Press any key to start the RAT Model simulation. ||"
+echo "||                                                                     ||"
+echo "========================================================================="
 
 # Wait for a keypress
 read -n 1 -s
 
-(cd ./rate-model && gnome-terminal -e "nrniv -python main.py" &
+gnome-terminal --working-directory "/home/ilyasse/Desktop/rat-model-launch" -- bash -c "nrniv -python main.py;"
 
 echo "Running RAT model... Ok".
